Fix print results for 32-bit system

+++ read_bw.c
@@ -340,7 +340,7 @@ static void print_report(struct perftest_parameters *user_param) {
 	tsize = user_param->duplex ? 2 : 1;
 	tsize = tsize * user_param->size;
 	
-	printf(REPORT_FMT,user_param->size,user_param->iters,!(user_param->noPeak) * tsize * cycles_to_units / opt_delta / 0x100000,
+	printf(REPORT_FMT,(unsigned long)user_param->size,user_param->iters,!(user_param->noPeak) * tsize * cycles_to_units / opt_delta / 0x100000,
 	       tsize * user_param->iters * cycles_to_units /(tcompleted[user_param->iters - 1] - tposted[0]) / 0x100000);
 }
